requirements.txt
alabaster==0.7.12
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.0
Django==2.1.1
django-cors-headers==3.1.0
django-model-utils==3.1.2
djangorestframework==3.9.3
docutils==0.15.2
idna==2.8
imageio==2.5.0
imagesize==1.1.0
Jinja2==2.10.1
joblib==0.13.2
JPype1==0.7.0
jsonpickle==1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
llvmlite==0.29.0
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.3
numba==0.45.1
numpy==1.16.2
packaging==19.1
pandas==0.24.2
Pillow==6.1.0
psycopg2-binary==2.8.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.2.1
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinx-autoapi==1.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.6
xlutils==2.0.0
xlwt==1.3.0
alabaster==0.7.12
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
Django==2.1.1
django-cors-headers==3.1.0
django-model-utils==3.1.2
djangorestframework==3.9.3
docutils==0.15.2
et-xmlfile==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
jdcal==1.4
Jinja2==2.10.1
jsonpickle==1.2
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
llvmlite==0.29.0
MarkupSafe==1.1.1
matplotlib==3.0.3
networkx==2.3
npm==0.1.1
numba==0.45.1
numpy==1.16.2
openpyxl==2.6.2
optional-django==0.1.0
packaging==19.1
pandas==0.24.2
Pillow==6.1.0
psycopg2-binary==2.8.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
scikit-image==0.15.0
scipy==1.2.1
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinx-autoapi==1.1.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.1.6
xlutils==2.0.0
xlwt==1.3.0
